+++ CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.3.0] - 2025-6-7
+
+### Added
+
++ Adapted to 1.21.70
++ Now freeCamera will keep player's skin
++ Added new command cfhud removeall
+
+### Fixed
+
++ add a feature that shortcut which include useless commonds will not be registed
+
 ## [2.2.1] - 2025-4-22
 
 ### Added